pyproject.toml
[project]
name = "gemini-ft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
